v1.7

+++ python/cepton_util/common.py
@@ -9,24 +9,35 @@ import sys
 import threading
 import time
 
+import numpy
 import serial
 import serial.tools.list_ports
 
 __all__ = [
     "add_execute_command_arguments",
     "add_io_path_arguments",
+    "add_simple_io_path_arguments",
     "ArgumentParserMixin",
     "create_directory",
     "execute_command",
+    "find_file_by_extension",
+    "find_file_by_name",
+    "find_file",
     "fix_path",
+    "get_day_str",
     "get_environment",
     "get_io_paths",
+    "get_sec_str",
+    "get_simple_io_paths",
     "get_timestamp_str",
     "get_timestamp",
+    "has_environment",
     "kill_background",
     "modify_path",
     "parse_execute_command_arguments",
+    "parse_list",
     "parse_time_hms",
+    "process_options",
     "remove_extension",
     "run_background",
     "set_extension",
@@ -35,14 +46,31 @@ __all__ = [
 ]
 
 
+def optional_function(func):
+    def wrapper(x, *args, **kwargs):
+        if x is None:
+            return None
+        return func(x, *args, **kwargs)
+    return wrapper
+
+
 def get_timestamp():
     return datetime.datetime.utcnow().timestamp()
 
 
+def get_day_str():
+    return datetime.datetime.now().strftime("%y%m%d")
+
+
+def get_sec_str():
+    return datetime.datetime.now().strftime("%H%M%S")
+
+
 def get_timestamp_str():
-    return datetime.datetime.now().strftime("%y%m%d_%H%M%S")
+    return datetime.datetime.now().strftime("{}_{}".format(get_day_str(), get_sec_str()))
 
 
+@optional_function
 def parse_time_hms(s):
     """Convert HH:MM:SS to seconds"""
     parts = [float(x) for x in s.split(":")]
@@ -52,14 +80,37 @@ def parse_time_hms(s):
     return sec
 
 
+@optional_function
+def parse_list(s, **kwargs):
+    options = {
+        "sep": ",",
+    }
+    options.update(kwargs)
+    return numpy.fromstring(s, **options)
+
+
+@optional_function
+def parse_enum(s, enum_type):
+    return enum_type[s.upper()]
+
+
 STRING_BOOL_LOOKUP = {
     "true": True,
     "false": False,
 }
 
 
-def get_environment(key):
-    s = os.environ[key]
+def has_environment(key):
+    return key in os.environ
+
+
+def get_environment(key, default=None):
+    try:
+        s = os.environ[key]
+    except:
+        if default is not None:
+            return default
+        raise
     if s.lower() in STRING_BOOL_LOOKUP:
         return STRING_BOOL_LOOKUP[s.lower()]
     return s
@@ -80,6 +131,9 @@ class BackgroundProcess(object):
         self.proc = subprocess.Popen(*args, **kwargs)
 
     def __del__(self):
+        self.close()
+
+    def close(self):
         try:
             if os.name == "nt":
                 os.kill(self.proc.pid, signal.CTRL_C_EVENT)
@@ -96,6 +150,9 @@ class BackgroundThread(object):
         self.thread = threading.Thread(*args, **kwargs)
 
     def __del__(self):
+        self.close()
+
+    def close(self):
         try:
             self.shutdown_event.set()
         except:
@@ -146,7 +203,7 @@ def execute_command(cmd_list, background=False, quiet=False, **kwargs):
     if background:
         proc = BackgroundProcess(cmd_list, **options)
         __local["procs"].append(proc)
-        return proc.proc
+        return proc
     else:
         subprocess.run(cmd_list, **options)
 
@@ -164,11 +221,40 @@ def parse_execute_command_arguments(args):
 # ------------------------------------------------------------------------------
 # Path
 # ------------------------------------------------------------------------------
+def find_file(func, path=None, depth=None):
+    if path is None:
+        path = os.getcwd()
+    level = 0
+    while os.path.isdir(path):
+        for name in os.listdir(path):
+            if func(name):
+                return os.path.join(path, name)
+
+        old_path = path
+        path = os.path.dirname(path)
+        if path == old_path:
+            break
+
+        level += 1
+        if (depth is not None) and (level >= depth):
+            break
+    return None
+
+
+def find_file_by_name(name, **kwargs):
+    return find_file(lambda x: x == name, **kwargs)
+
+
+def find_file_by_extension(ext, **kwargs):
+    return find_file(lambda x: x.endswith(ext), **kwargs)
+
+
 def create_directory(path):
     if not os.path.exists(path):
         os.makedirs(path)
 
 
+@optional_function
 def fix_path(path):
     path = os.path.expanduser(path)
     path = os.path.expandvars(path)
@@ -177,14 +263,17 @@ def fix_path(path):
     return path
 
 
+@optional_function
 def remove_extension(path):
     return os.path.splitext(path)[0]
 
 
+@optional_function
 def set_extension(path, new_ext):
     return os.path.splitext(path)[0] + new_ext
 
 
+@optional_function
 def modify_path(path, new_ext=None, prefix="", postfix=""):
     head, basename = os.path.split(path)
     name, ext = os.path.splitext(path)
@@ -195,29 +284,48 @@ def modify_path(path, new_ext=None, prefix="", postfix=""):
     return path
 
 
-def add_io_path_arguments(parser, prefix="", postfix=""):
+def add_simple_io_path_arguments(parser):
     parser.add_argument("input")
+    parser.add_argument("--dry_run", action="store_true")
     parser.add_argument("-o", "--output")
+
+
+def get_simple_io_paths(args, output_name):
+    input_path = fix_path(args.input)
+    if args.output is not None:
+        output_path = fix_path(args.output)
+    else:
+        dir_path = os.path.dirname(input_path)
+        output_path = os.path.join(dir_path, output_name)
+    return input_path, output_path
+
+
+def add_io_path_arguments(parser, prefix="", postfix=""):
+    add_simple_io_path_arguments(parser)
     parser.add_argument("--inplace", action="store_true")
     parser.add_argument("--prefix", default=prefix)
     parser.add_argument("--postfix", default=postfix)
 
 
-def get_io_paths(args, output_ext=None, **kwargs):
+def get_io_paths(args, output_ext=None, output_name=None):
     input_path = fix_path(args.input)
-    if args.inplace:
-        output_path = input_path
-    elif args.output is None:
-        options = {
-            "postfix": args.postfix,
-            "prefix": args.prefix,
-        }
-        output_path = modify_path(input_path, **options)
-    else:
+    if args.output is not None:
         output_path = fix_path(args.output)
+    elif args.inplace:
+        output_path = input_path
+    else:
+        if output_name is None:
+            options = {
+                "postfix": args.postfix,
+                "prefix": args.prefix,
+            }
+            output_path = modify_path(input_path, **options)
+        else:
+            dir_path = os.path.dirname(input_path)
+            output_path = os.path.join(dir_path, output_name)
     if output_ext is not None:
         output_path = modify_path(output_path, new_ext=output_ext)
-    if input_path == output_path:
+    if (not args.dry_run) and (input_path == output_path):
         input_path = modify_path(input_path, postfix="_old")
         shutil.move(output_path, input_path)
     return input_path, output_path
@@ -241,3 +349,7 @@ class ArgumentParserMixin(object):
         options = cls.parse_arguments(parser_args)
         options.update(kwargs)
         return cls(*args, **options)
+
+
+def process_options(options):
+    return {key: value for key, value in options.items() if value is not None}
